+++ pkg/resources/database_acceptance_test.go
@@ -1,14 +1,21 @@
 package resources_test
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"strings"
 	"testing"
 
 	acc "github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+	"github.com/hashicorp/terraform-plugin-testing/config"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+	"github.com/hashicorp/terraform-plugin-testing/tfversion"
+	"github.com/stretchr/testify/require"
 )
 
 func TestAcc_DatabaseWithUnderscore(t *testing.T) {
@@ -83,6 +90,50 @@ func TestAcc_Database(t *testing.T) {
 	})
 }
 
+func TestAcc_DatabaseRemovedOutsideOfTerraform(t *testing.T) {
+	id := generateUnsafeExecuteTestDatabaseName(t)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: acc.TestAccProtoV6ProviderFactories,
+		PreCheck:                 func() { acc.TestAccPreCheck(t) },
+		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
+			tfversion.RequireAbove(tfversion.Version1_5_0),
+		},
+		CheckDestroy: testAccCheckDatabaseExistence(t, id, false),
+		Steps: []resource.TestStep{
+			{
+				ConfigDirectory: config.TestNameDirectory(),
+				ConfigVariables: map[string]config.Variable{
+					"db": config.StringVariable(id),
+				},
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{plancheck.ExpectNonEmptyPlan()},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("snowflake_database.db", "name", id),
+					resource.TestCheckResourceAttr("snowflake_database.db", "comment", "test comment"),
+					testAccCheckDatabaseExistence(t, id, true),
+				),
+			},
+			{
+				PreConfig:       func() { dropDatabaseOutsideTerraform(t, id) },
+				ConfigDirectory: config.TestNameDirectory(),
+				ConfigVariables: map[string]config.Variable{
+					"db": config.StringVariable(id),
+				},
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{plancheck.ExpectNonEmptyPlan()},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("snowflake_database.db", "name", id),
+					resource.TestCheckResourceAttr("snowflake_database.db", "comment", "test comment"),
+					testAccCheckDatabaseExistence(t, id, true),
+				),
+			},
+		},
+	})
+}
+
 func dbConfig(prefix string) string {
 	s := `
 resource "snowflake_database" "db" {
@@ -102,4 +153,36 @@ resource "snowflake_database" "db" {
 }
 `
 	return fmt.Sprintf(s, prefix)
+}
+
+func dropDatabaseOutsideTerraform(t *testing.T, id string) {
+	t.Helper()
+
+	client, err := sdk.NewDefaultClient()
+	require.NoError(t, err)
+	ctx := context.Background()
+
+	err = client.Databases.Drop(ctx, sdk.NewAccountObjectIdentifier(id), &sdk.DropDatabaseOptions{})
+	require.NoError(t, err)
+}
+
+func testAccCheckDatabaseExistence(t *testing.T, id string, shouldExist bool) func(state *terraform.State) error {
+	t.Helper()
+	return func(state *terraform.State) error {
+		client, err := sdk.NewDefaultClient()
+		require.NoError(t, err)
+		ctx := context.Background()
+
+		_, err = client.Databases.ShowByID(ctx, sdk.NewAccountObjectIdentifier(id))
+		if shouldExist {
+			if err != nil {
+				return fmt.Errorf("error while retrieving database %s, err = %w", id, err)
+			}
+		} else {
+			if err == nil {
+				return fmt.Errorf("database %v still exists", id)
+			}
+		}
+		return nil
+	}
 }