+++ pkg/sdk/client.go
@@ -54,6 +54,7 @@ type Client struct {
 	Shares           Shares
 	Stages           Stages
 	Streams          Streams
+	Tables           Tables
 	Tags             Tags
 	Tasks            Tasks
 	Users            Users
@@ -185,6 +186,7 @@ func (c *Client) initialize() {
 	c.Stages = &stages{client: c}
 	c.Streams = &streams{client: c}
 	c.SystemFunctions = &systemFunctions{client: c}
+	c.Tables = &tables{client: c}
 	c.Tags = &tags{client: c}
 	c.Tasks = &tasks{client: c}
 	c.Users = &users{client: c}